requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
qdrant-client==1.9.2
sentence-transformers==3.0.1
numpy==1.26.4
redis==5.0.8
rq==1.16.2
requests==2.32.3
python-dotenv==1.0.1
transformers==4.44.2
streamlit==1.37.1
ragas==0.1.9
pytest==8.3.2
httpx==0.27.2
